// frontend/src/services/OpenApi/models/PagedModelOpeningDetailsActivitiesActivitiesDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OpeningDetailsActivitiesActivitiesDto } from './OpeningDetailsActivitiesActivitiesDto';
import type { PageMetadata } from './PageMetadata';
export type PagedModelOpeningDetailsActivitiesActivitiesDto = {
    content?: Array<OpeningDetailsActivitiesActivitiesDto>;
    page?: PageMetadata;
};